requirements.txt
Flask
ultralytics
torch
torchvision
opencv-python
Pillow
numpy
requests
python-dotenv
flask_socketio
flask_cors
eventlet
